+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### [0.25.0](https://kaos.sh/pachca/0.25.0)
+
+- Added helper `User.Mention`
+- Added method `UpdateAvatar`
+- Added method `DeleteAvatar`
+- Added method `UpdateUserAvatar`
+- Added method `DeleteUserAvatar`
+
 ### [0.24.0](https://kaos.sh/pachca/0.24.0)
 
 - Improved API errors handling